@@ -1,48 +1,35 @@
 <script>
     import {ComponentBox,PackageComponent, GeneralInfo} from "$lib/index.js";
-	export let data;
+    import {WinC,Button} from "$lib";
 	// import '../../styles/overview.css'
     // import {MobilityOptions,GeneralInfo, ,ComponentBox} from '$lib';
    
 
-  export let num = 0;
+  /** @type {{data: any, num?: number}} */
+  let { data, num = 0 } = $props();
 </script>
 
 <main>
    
-            <ComponentBox
-                heading="Mobility options"
-                context1=""
-                context2=""
-                display="flex"
-                titleAlign="center"
-                class="top-c"
-            ></ComponentBox>
 
-    <ComponentBox
-    heading=""
-    context1=""
-    context2=""
-    display="flex"
-    titleAlign=""
+    <WinC
+    role="child"
+    context=""
     class="box2"
     >
         <PackageComponent {data} />
-    
-    </ComponentBox>
+        <p>cards</p>
+    </WinC>
 
-    <ComponentBox
-    heading=""
-		context1=""
-		context2=""
-		display="flex"
-		titleAlign=""
+    <WinC
+        role="child"
+		context=""
 		class="box1"
     >
         <GeneralInfo 
         {num}/>
 
-    </ComponentBox>
+    </WinC>
 </main>
 
 <style>
@@ -56,5 +43,7 @@
         display: flex;
         flex-direction: column;
         gap: 2em;
+
+       
     }
 </style>
